@@ -16,19 +16,22 @@
     "@crxjs/vite-plugin": "2.0.0-beta.19",
     "@next/eslint-plugin-next": "14.0.0",
     "@types/chrome": "0.0.248",
-    "@types/eslint": "8.44.6",
-    "@types/node": "20.8.9",
-    "@types/react": "18.2.33",
-    "@types/react-dom": "18.2.14",
-    "@vitejs/plugin-react": "4.1.0",
-    "eslint-config-upleveled": "5.1.1",
-    "typescript": "5.2.2",
+    "@types/eslint": "8.44.7",
+    "@types/node": "20.10.0",
+    "@types/react": "18.2.38",
+    "@types/react-dom": "18.2.17",
+    "@typescript-eslint/utils": "^6.12.0",
+    "@vitejs/plugin-react": "4.2.0",
+    "eslint-config-upleveled": "7.3.3",
+    "globals": "^13.23.0",
+    "prettier": "^3.1.0",
+    "typescript": "5.3.2",
     "vite": "5.0.0"
   },
   "pnpm": {
     "patchedDependencies": {
-      "@vitejs/plugin-react@4.1.0": "patches/@vitejs__plugin-react@4.1.0.patch",
-      "@crxjs/vite-plugin@2.0.0-beta.19": "patches/@crxjs__vite-plugin@2.0.0-beta.19.patch"
+      "@crxjs/vite-plugin@2.0.0-beta.19": "patches/@crxjs__vite-plugin@2.0.0-beta.19.patch",
+      "@vitejs/plugin-react@4.2.0": "patches/@vitejs__plugin-react@4.2.0.patch"
     }
   }
 }
