@@ -4,4 +4,4 @@ const config = {
   trailingComma: 'all',
 };
 
-module.exports = config;
+export default config;
